refactor: split chatmate_test.go and CONTRIBUTING.md for 300-line compliance; make yaml.v3 direct dep
